Fix bugs when editing commands and actions
The command and action forms were making changes even if Cancel was
clicked, in some cases causing data to be lost. Now they only make
changes if OK is clicked.

Set the DialogResult value of OK and Cancel buttons so that forms close
on click without a click event handler.

Remove the click event handlers on OK and Cancel.

Adjust the code which opened each form to check ShowDialog() for its
DialogResult.OK return value and only make changes in that case.

Change both forms to take a copy of the data they're working with rather
than trying to modify the data in place. This especially helps when
actions are added/removed/changed but then all changes to the command
are cancelled; unless copies are taken, this doesn't work, and changes
were still happening.